+++ src/components/Router.js
@@ -1,10 +1,12 @@
-import React, { PropTypes } from 'react';
-import { LobbyContainer as lobby } from '../containers/LobbyContainer';
-import { ChatContainer as chat } from '../containers/ChatContainer';
+import React, {PropTypes} from 'react';
+import {LobbyContainer as lobby} from '../containers/LobbyContainer';
+import {ChatContainer as chat} from '../containers/ChatContainer';
+import {Guest as guest} from './guest';
 
 const Handlers = {
     lobby,
     chat,
+    guest,
 };
 
 export const Router = (props) => {