+++ src/router/Routes.js
@@ -6,6 +6,9 @@ export const Routes = [
         onEnter: (dispatch) => () => dispatch(leave()),
     },
     {
+        path: '/guest', route: 'guest',
+    },
+    {
         path: '/:roomId', route: 'chat',
         onEnter: (dispatch) => ({roomId}) => dispatch(join(roomId)),
     },