+++ src/server/app.js
@@ -1,13 +1,15 @@
 import { urlencoded } from 'body-parser';
 import config from 'config';
 import express from 'express';
-import { getLogger } from 'log4js';
+import {getLogger} from 'log4js';
 import Livereload from 'connect-livereload';
-import { Message } from './models/message';
-import { Room, PASSWORD_INCORRECT } from './models/room';
-import { knex, exists } from './knex';
-import { session } from './session';
-import { getUser } from './user';
+import {Message} from './models/message';
+import {NOT_FOUND} from './models/model';
+import {Room, PASSWORD_INCORRECT} from './models/room';
+import {User} from './models/user';
+import {knex, exists} from './knex';
+import {session} from './session';
+import {getUser} from './user';
 
 const browser = config.get('browser');
 const logger = getLogger('[app]');
@@ -42,7 +44,7 @@ const staticView = (req, res, next) => {
             .join(req.params.roomId, req.body && req.body.password)
             .then((room) => ({ room, user }))
         )
-        .then(({ room, user }) => Message
+        .then(({room, user}) => Message
             .findAll(room.id, user.id)
             .then((messages) => ({
                 ...room,
@@ -57,6 +59,8 @@ const staticView = (req, res, next) => {
         .catch((e) => {
             if (e === PASSWORD_INCORRECT) {
                 return res.redirect(`/view/${req.params.roomId}/password`);
+            } else if (e === NOT_FOUND) {
+                return res.redirect('/guest');
             }
             logger.error(e);
 
@@ -64,27 +68,68 @@ const staticView = (req, res, next) => {
         });
 };
 app.get('/view/:roomId', staticView);
-app.post('/view/:roomId', urlencoded({ extended: false }), staticView);
+app.post('/view/:roomId', urlencoded({extended: false}), staticView);
 app.get('/view/:roomId/password', (req, res) => {
     res.render('static-password', {
         id: req.params.roomId,
     });
 });
 
+const renderIndex = (res, user = null) =>
+    res.render('index', {
+                config: browser,
+                script: process.env.NODE_ENV === 'production'
+                    ? 'js/browser.min.js'
+                    : 'js/browser.js',
+                ga: config.has('ga') &&
+                    `GA_CONFIG = ${JSON.stringify(config.get('ga'))};`,
+                user,
+        }
+    );
 
-app.get(['/', '/:roomId'], (req, res) => {
-    getUser(req.session)
-        .then((user) => res.render('index', {
-            config: browser,
-            script: process.env.NODE_ENV === 'production'
-                ? 'js/browser.min.js'
-                : 'js/browser.js',
-            ga: config.has('ga') &&
-                `GA_CONFIG = ${JSON.stringify(config.get('ga'))};`,
-            user,
-        }))
+app.get('/logout', (req, res, next) => {
+    if (!config.get('app.guest')) return next();
+
+    req.session.guest = null;
+
+    return res.redirect('/');
+});
+app.get('/guest', (req, res, next) => {
+    if (!config.get('app.guest')) return next();
+
+    return renderIndex(res);
+});
+app.post('/guest', urlencoded({extended: false}), (req, res, next) => {
+    if (!config.get('app.guest')) return next();
+
+
+    const guest = {
+        id: req.body.id,
+        name: req.body.name,
+    };
+
+    return User.find('id', guest.id)
+        .then(() => {
+            next();
+        })
         .catch((e) => {
-            logger.error(e);
-            res.sendStatus(401);
+            if (e !== NOT_FOUND) return next();
+
+            req.session.guest = guest;
+
+            return res.redirect('/');
         });
 });
+
+app.get(['/', '/:roomId'], (req, res, next) =>
+    getUser(req.session)
+        .then((user) => renderIndex(res, user))
+        .catch((e) => {
+            if (config.get('app.guest') && e === NOT_FOUND) {
+                return res.redirect('/guest');
+            }
+
+            logger.error(e);
+            next(e);
+        })
+);