+++ src/server/models/__tests__/test-user.js
@@ -1,0 +1,10 @@
+describe('User', () => {
+    const {Model} = require('../model');
+
+    jest.dontMock('../user');
+    const {User} = require('../user');
+
+    it('inherits Model', () => {
+        expect(Model).toBeCalledWith('users');
+    });
+});