+++ src/server/models/user.js
@@ -1,0 +1,9 @@
+import {Model} from './model';
+
+export class UserModel extends Model {
+    constructor() {
+        super('users');
+    }
+}
+
+export const User = new UserModel();