+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Guest login with any user id.
+
 ## [1.2.1] - 2016-03-19
 ### Added
 - Room state. ([#45](https://github.com/ukatama/nekochat/issues/45))